Split file store backends
