+++ src/__test__/fileStore.test.js
@@ -6,6 +6,7 @@ import {
   DiskFileBackend,
   S3FileBackend,
 } from '../fileStore/backends';
+
 import path from 'path';
 import fs from 'fs';
 import tempy from 'tempy';
